requirements.txt
# requirements.txt
streamlit==1.31.0
langchain==0.1.13
langchain-community==0.0.29
langchain-openai==0.0.8
openai==1.14.0
google-api-python-client==2.121.0
google-auth==2.28.2
requests==2.31.0
beautifulsoup4==4.12.3